python/requirement.txt
tensorflow==2.3.4
tensorflow-gpu==2.3.4
transformers==3.3.1
tf2onnx==1.12.0
onnxruntime==1.12.1
pandas
seqeval
fastprogress
attrdict
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cu101
torchvision==0.7.0+cu101


